backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
requests==2.31.0
PyJWT==2.8.0
google-auth==2.25.0 